# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAliveResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"payload": {
						"properties": {
							"vip": {
								"type": "boolean"
							},
							"discount": {
								"type": "number",
								"format": "double"
							},
							"group": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"address": {
								"type": "string"
							},
							"isEmployee": {
								"type": "boolean"
							},
							"dni": {
								"type": "string"
							},
							"createdOn": {
								"type": "string",
								"format": "date-time"
							},
							"profilePicture": {
								"type": "string"
							},
							"isProvider": {
								"type": "boolean"
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"token": {
								"type": "string"
							}
						},
						"required": [
							"isProvider",
							"email",
							"name",
							"token"
						],
						"type": "object"
					}
				},
				"required": [
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserRegister": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"dni": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserProfileUpdate": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"dni": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"googleUser": {
				"properties": {
					"picture": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object"
			},
			"IUserAuthenticated": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"dni": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"isProvider": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password",
					"isProvider"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticatedResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/IUserAuthenticated"
					},
					"token": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"IUserAuthenticate": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISupplier": {
				"properties": {
					"name": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"cuit": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"primaryColor": {
						"type": "string"
					},
					"secondaryColor": {
						"type": "string"
					},
					"coverPhoto": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isProvider": {
						"type": "boolean"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"businessName",
					"cuit",
					"domain",
					"address",
					"phone",
					"category",
					"email",
					"primaryColor",
					"secondaryColor",
					"password",
					"isProvider",
					"createdOn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticatedSupplierResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"supplier": {
						"$ref": "#/components/schemas/ISupplier"
					},
					"token": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"AuthenticateSupplier": {
				"properties": {
					"password": {
						"type": "string"
					},
					"cuit": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"cuit"
				],
				"type": "object"
			},
			"ISupplierResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"token": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/ISupplier"
					}
				},
				"required": [
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISupplierRegister": {
				"properties": {
					"name": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"cuit": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"primaryColor": {
						"type": "string"
					},
					"secondaryColor": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"businessName",
					"cuit",
					"domain",
					"address",
					"phone",
					"category",
					"email",
					"primaryColor",
					"secondaryColor",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISupplierUpdate": {
				"properties": {
					"name": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"cuit": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"primaryColor": {
						"type": "string"
					},
					"secondaryColor": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"cuit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"deliveryStatus": {
				"type": "string",
				"enum": [
					"waiting_for_package",
					"waiting_for_robot",
					"on_transit",
					"delivered"
				]
			},
			"IOrder": {
				"properties": {
					"productName": {
						"type": "string"
					},
					"productPrice": {
						"type": "number",
						"format": "double"
					},
					"productQuantity": {
						"type": "number",
						"format": "double"
					},
					"marketplace": {
						"type": "string"
					},
					"cuit": {
						"type": "string"
					},
					"purchaseId": {
						"type": "string"
					},
					"userEmail": {
						"type": "string"
					},
					"userDni": {
						"type": "string"
					},
					"deliveryStatus": {
						"$ref": "#/components/schemas/deliveryStatus"
					},
					"orderDate": {
						"type": "string",
						"format": "date-time"
					},
					"deliveryDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"productName",
					"productPrice",
					"productQuantity",
					"marketplace",
					"cuit",
					"purchaseId",
					"userEmail",
					"userDni"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT Bearer token **_without_** the 'Bearer ' prefix"
			}
		}
	},
	"info": {
		"title": "deliverar-users-service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/alive": {
			"get": {
				"operationId": "Alive",
				"responses": {
					"200": {
						"description": "A message with the API health status.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAliveResponse"
								}
							}
						}
					}
				},
				"description": "Get API Health Status.",
				"tags": [
					"API Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "The user data.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"description": "Register a new user.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserRegister"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "The user data.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"description": "Update a new user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserProfileUpdate"
							}
						}
					}
				}
			}
		},
		"/api/user/profilePicture/{email}": {
			"put": {
				"operationId": "UpdateProfilePicture",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"profilePicture": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"profilePicture"
								]
							}
						}
					}
				}
			}
		},
		"/api/user/{email}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"description": "Deletes User Account.",
				"tags": [
					"User"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The user email.",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user/handleGoogleLogin": {
			"post": {
				"operationId": "HandleGoogleLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"description": "Get User Data.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/googleUser"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "The user data.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthenticatedResponse"
								}
							}
						}
					}
				},
				"description": "Authenticate a user.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserAuthenticate"
							}
						}
					}
				}
			}
		},
		"/api/auth/login-supplier": {
			"post": {
				"operationId": "LoginSupplier",
				"responses": {
					"200": {
						"description": "The supplier data.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthenticatedSupplierResponse"
								}
							}
						}
					}
				},
				"description": "Authenticate a supplier.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticateSupplier"
							}
						}
					}
				}
			}
		},
		"/api/auth/password-recovery/{email}": {
			"post": {
				"operationId": "Recovery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthenticatedResponse"
								}
							}
						}
					}
				},
				"description": "Initiates User Password Recovery Process.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user email.",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/supplier": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "The supplier data with status code.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Register a new supplier.",
				"tags": [
					"Supplier"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The supplier registration data.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISupplierRegister",
								"description": "The supplier registration data."
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "The updated supplier data with status code.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Updates supplier information.",
				"tags": [
					"Supplier"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The supplier update data.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISupplierUpdate",
								"description": "The supplier update data."
							}
						}
					}
				}
			}
		},
		"/api/supplier/logo/{cuit}": {
			"put": {
				"operationId": "UpdateLogo",
				"responses": {
					"200": {
						"description": "The updated supplier data with status code.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Updates supplier logo.",
				"tags": [
					"Supplier"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The supplier CUIT.",
						"in": "path",
						"name": "cuit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"logo": {
										"type": "string",
										"format": "binary",
										"description": "The supplier's new logo."
									}
								},
								"required": [
									"logo"
								]
							}
						}
					}
				}
			}
		},
		"/api/supplier/coverPhoto/{cuit}": {
			"put": {
				"operationId": "UpdateCoverPhoto",
				"responses": {
					"200": {
						"description": "The updated supplier data with status code.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Updates supplier cover photo.",
				"tags": [
					"Supplier"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The supplier CUIT.",
						"in": "path",
						"name": "cuit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"coverPhoto": {
										"type": "string",
										"format": "binary",
										"description": "The supplier's new cover photo."
									}
								},
								"required": [
									"coverPhoto"
								]
							}
						}
					}
				}
			}
		},
		"/api/supplier/{cuit}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Status code of operation.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Deletes supplier account.",
				"tags": [
					"Supplier"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The supplier cuit.",
						"in": "path",
						"name": "cuit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "The supplier data with status code.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISupplierResponse"
								}
							}
						}
					}
				},
				"description": "Retrieves supplier information.",
				"tags": [
					"Supplier"
				],
				"security": [],
				"parameters": [
					{
						"description": "The supplier cuit.",
						"in": "path",
						"name": "cuit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/order/{email}": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "The user orders.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all orders for given user.",
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user email.",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/order/supplier/{cuit}": {
			"get": {
				"operationId": "GetOrdersSupplierCuit",
				"responses": {
					"200": {
						"description": "The supplier order.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all orders for given Supplier.",
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"description": "The supplier CUIT.",
						"in": "path",
						"name": "cuit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}